+++ src/MiNET/MiNET/Entities/Hostile/Witch.cs
@@ -29,7 +29,7 @@ namespace MiNET.Entities.Hostile
 {
 	public class Witch : HostileMob
 	{
-		public Witch(Level level) : base((int) EntityType.Witch, level)
+		public Witch(Level level) : base(EntityType.Witch, level)
 		{
 			Width = Length = 0.6;
 			Height = 1.8;
