+++ src/MiNET/MiNET/Entities/Hostile/WitherSkeleton.cs
@@ -29,7 +29,7 @@ namespace MiNET.Entities.Hostile
 {
 	public class WitherSkeleton : HostileMob, IAgeable
 	{
-		public WitherSkeleton(Level level) : base((int) EntityType.WitherSkeleton, level)
+		public WitherSkeleton(Level level) : base(EntityType.WitherSkeleton, level)
 		{
 			Width = Length = 0.7;
 			Height = 2.4;
